Jmol.___JmolVersion="14.3.16_2015.09.28"
new feature: array.sort("key")
 -- sorts an array of associative arrays by the specified associative array key.
 -- example:
 
	load sf6.smol
	x= _M.moData.mos.select("energy,index where energy<0")
	x.sort("energy").reverse
	print x

	  {
	    "energy"  :  -0.67001194
	    "index"  :  35
	   }
	  {
	    "energy"  :  -0.67001194
	    "index"  :  34
	   }
	  {
	    "energy"  :  -0.67001194
	    "index"  :  33
	   }
	  {
	    "energy"  :  -0.70143324
	    "index"  :  32
	   }
	   ...
	   	
new feature: JmolSQL xxxx.select("<keys> WHEREIN <clause>")
  -- replaces ** idea
  -- assumes values of xxxx are themselves associative arrays
  -- indicates that the clause should be applied only to the values for the selected keys
  -- returns an associative array that holds only those keys for which the clause is true *for them*
  -- works also for array xxxx where each element of xxxx is an associative array, 
     in which case it returns a subarray of xxxx involving the matching values.  
  -- example:
  
   abc = [ A:[b:1], B:[b:2], C:[b:3, d:50], AA:[b:4] ]
   print abc.select("A* wherein b>2").format("JSON")
   
     { "AA": { "b": 4 } }
     
   Compare this to the following, where we are are checking for validity of abc itself:
   
   abc = [ A:[b:1], B:[b:2], C:[b:3, d:50], AA:[b:4] ]
   print abc.select("A* where B.b=2").format("JSON")
   
     { "A": { "b": 1 },"AA": { "b": 4 } }

   
   abc2 = [ 
      [ A:[b:1], B:[b:2], C:[b:3, d:50], AA:[b:4] ],
      [ A:[b:11], B:[b:22], C:[b:33, d:5050], AA:[b:40] ]
      ]
   print abc2.select("A* wherein b>2").format("JSON")
    
   [ { "AA": { "b": 4 } },{ "A": { "b": 11 },"AA": { "b": 40 } } ]      

new feature: JmolSQL xxxx.select("(<keys>) WHERE/WHEREIN <clause>")
  -- parentheses around keys replaces ";" idea
  -- indicates to return an array of values rather than a key/value pair
  -- works also for array xxxx where each element of xxxx is an array or associative array, 
     in which case it returns an array of matching values, without keys.  
  -- example:

   load =1ehz/dssr
   print _M.dssr.hbonds.select("distance where res_long like '*|C|72|*'")
	  {
	    "distance"  :  2.832
	   }
	  {
	    "distance"  :  2.879
	   }
	  {
	    "distance"  :  2.838
	   }

	print _M.dssr.hbonds.select("(distance) where res_long like '*|C|72|*'")
	
		2.832
		2.879
		2.838

	print _M.dssr.hbonds.select("(distance) where res_long like '*|C|72|*'").average

		2.8496666
       

So...

JmolSQL

There are three parts to JmolSQL, object, keys, and an optional WHERE or WHERIN phrase:

object.SELECT("keys WHERE/WHEREIN phrase")

The object can be either an associative array [key1:value1, key2:value2, key3:value3] 
or an array of associative arrays, usually with the same set of keys.

Associative Arrays

When the top-level array is an associative array, select() can be used to select 
out subsets of the array, either as a single associative array or as an array of values.

abc.select("A,B")

returns a subset of abc. Wild cards can be interspersed with additional keys, 
for example, "a*,b" or "*_id". In each such case, the case-sensitive LIKE operation is used to match keys.

abc = [ A:[b:1],B:[b:2],AA:[b:3, d:50] ]
print  abc.select("A").format("JSON")

{ "A": { "b": 1 } }

abc = [ A:[b:1],B:[b:2],AA:[b:3, d:50] ]
print  abc.select("A*").format("JSON")

{ "A": { "b": 1 },"AA": { "b": 3,"d": 50 } }

abc.select("(...)")

Using parentheses around the list of keys delivers a list of values of b for only the the subset of xyz for which a=1:

abc = [ A:[b:1],B:[b:2],AA:[b:3, d:50] ]
print  abc.select("(A,B)").format("JSON")

[ { "b": 2 },{ "b": 1 } ]

Arrays of Associative Arrays

Generally we assume here that the elements of the array are associative arrays, all with the 
same set of keys. This is the essence of a database.  Whether or not the associative 
arrays have the same keys is not important for Jmol, but is typical for a database.

For example:

xyz = [ [a:1,b:11],[a:2,b:22],[a:3,b:33] ]


xyz.select("b")

Creates the sublist of associative arrays having the selected subset of keys:

xyz = [ [a:1,b:11],[a:2,b:22],[a:3,b:33] ]
print xyz.select("b").format("JSON")

[ { "b": 11 },{ "b": 22 },{ "b": 33 } ]

xyz.select("(...)")

Creates a list of only the b values of each of x, y, and z, in order:

xyz = [ [a:1,b:11],[a:2,b:22],[a:3,b:33] ]
print xyz.select("(b)").format("JSON")

[ 11,22,33 ]

xyz = [ [a:1,b:11],[a:2,b:22],[a:3,b:33] ]
print xyz.select("(a,b)").format("JSON")

[ 11,1,22,2,33,3 ]

The assumption when using (keys) is that you want to know all these values, 
but you don't care what order they are in (because that will not be predictable) 
and you don't care about their exact context. An example is a list of bonds for 
which we just want to know all the atoms involved, but the atoms are listed 
under "atom1" and "atom2" in each bond array.

load =1ehz/dssr
select on @{_M.dssr.hbonds.select("(atom1_id,atom2_id)")}

206 atoms selected

Using WHERE

WHERE is used to select a subset of the elements of an array based on specific
key-value relationships.

xyz.select("... WHERE ...")

Delivers all key/value pairs in the subset of xyz element associative arrays 
for which the WHERE clause is true for that element. 

xyz = [ [a:1,b:11],[a:2,b:22],[a:3,b:33] ]
print xyz.select("* where a<3 and b<20").format("JSON")

[ { "b": 11,"a": 1 } ]

xyz.select("(...) where ...")

Using parentheses around the list of keys delivers a list of 
values of for only the the subset of xyz for which the WHERE clause is true:

xyz = [ [a:1,b:11],[a:2,b:22],[a:3,b:33] ]
print xyz.select("(b) where a>1").format("JSON")

[ 22,33 ]

Note that "... where..." will "drill down" through arrays of arrays to find the 
targeted associative array elements, producing a flat array of those objects:

xyz = [ [[a:1,b:11], [a:0,b:0]],[[[a:2,b:22]]],[[a:3,b:33,aa:44]] ]
print xyz.select("a* where a>0").format("JSON")

[ { "a": 1 },{ "a": 2 },{ "a": 3,"aa": 44 } ]

xyz = [ [[a:1,b:11], [a:0,b:0]],[[[a:2,b:22]]],[[a:3,b:33,aa:44]] ]
print xyz.select("(b) where a>0").format("JSON")

[ 11,22,33 ]
 
WHERE vs. WHEREIN

Starting with Jmol 14.4, JmolSQL includes an additional option, WHEREIN. This option 
allows selecting specific key/value pairs for which the value is itself an associative 
array, and *that array* has a specific set of key/value relationships. Thus, the
clause is checked one level deeper in the structure.

For example, given the associative array

abc = [key_1:[type:"a", i:1],key_2:[type:"b", i:2],key_3:[type:"a", i:3]]

we can select out only those keys for which type='a':

abc = [key_1:[type:"a", i:1],key_2:[type:"b", i:2],key_3:[type:"a", i:3]]
print abc.select("* WHEREIN type='a'").format("JSON");

{ "key_3": { "i": 3,"type": "a" },"key_1": { "i": 1,"type": "a" } }

All of the options that involve WHERE also apply to WHEREIN. For example,
multiple keys can be specified, and keys can be surrounded by parentheses 
to return just the values:

abc = [key_1:[type:"a", i:1],key_2:[type:"b", i:2],key_3:[type:"a", i:3]]
print abc.select("(key_1,key2) WHEREIN type='a'").format("JSON");

[ { "i": 1,"type": "a" } ]

In addition, WHEREIN can be applied to arrays as well as associative arrays. 
In this case, the WHEREIN phrase applies to the elements of that array, which
are assumed to be associative arrays. For example, we can get a list of just
the orbitals that are of a given symmetry:

load http://chemapps.stolaf.edu/jmol/jsmol/data/no2_nbo.log.gz 2 filter "alpha"
x= _M.moData.select("mos wherein occupancy>0 and symmetry like '(B2)*' ")
print x.select("(index)").format("JSON")

[ 1,5,8,10 ]



git-svn-id: http://svn.code.sf.net/p/jmol/code/trunk/Jmol@20797 62be27d7-c70d-0410-b753-bb59b6ba1b27